Refactor MidiEndpointConnection into multiple source files
