+++ src/api/Client/Midi2Client/MidiEndpointConnection.cpp
@@ -16,73 +16,6 @@
 namespace winrt::Windows::Devices::Midi2::implementation
 {
     _Use_decl_annotations_
-    midi2::MidiSendMessageResult MidiEndpointConnection::SendMessageResultFromHRESULT(HRESULT hr)
-    {
-        midi2::MidiSendMessageResult result{ 0 };
-
-        if (SUCCEEDED(hr))
-        {
-            result |= midi2::MidiSendMessageResult::Succeeded;
-        }
-        else
-        {
-            result |= midi2::MidiSendMessageResult::Failed;
-        }
-
-        switch (hr)
-        {
-        //case HR_S_MIDI_SENDMSG_IMMEDIATE:
-        //    result |= midi2::MidiSendMessageResult::SentImmediately;
-        //    break;
-
-        //case HR_S_MIDI_SENDMSG_SCHEDULED:
-        //    result |= midi2::MidiSendMessageResult::Scheduled;
-        //    break;
-
-        //case HR_S_MIDI_SENDMSG_SYSEX_PARKED:
-        //    result |= midi2::MidiSendMessageResult::SystemExclusiveParked;
-        //    break;
-
-        //case HR_E_MIDI_SENDMSG_BUFFER_FULL:
-        //    result |= midi2::MidiSendMessageResult::BufferFull;
-        //    break;
-
-        //case HR_E_MIDI_SENDMSG_SCHEDULER_QUEUE_FULL:
-        //    result |= midi2::MidiSendMessageResult::SchedulerQueueFull;
-        //    break;
-
-        //case HR_E_MIDI_SENDMSG_INVALID_MESSAGE:
-        //    result |= midi2::MidiSendMessageResult::InvalidMessageOther;
-        //    break;
-
-        case HRESULT_FROM_WIN32(ERROR_INSUFFICIENT_BUFFER):
-            result |= midi2::MidiSendMessageResult::BufferFull;
-            break;
-
-        default:
-            result |= midi2::MidiSendMessageResult::Other;
-            break;
-        }
-
-        return result;
-    }
-
-
-    _Use_decl_annotations_
-    bool MidiEndpointConnection::ValidateUmp(uint32_t word0, uint8_t wordCount) noexcept
-    {
-        if (!internal::IsValidSingleUmpWordCount(wordCount))
-            return false;
-
-        if (internal::GetUmpLengthInMidiWordsFromFirstWord(word0) != wordCount)
-            return false;
-
-        return true;
-    }
-
-
-
-    _Use_decl_annotations_
     HRESULT MidiEndpointConnection::Callback(PVOID data, UINT size, LONGLONG timestamp, LONGLONG)
     {
         internal::LogInfo(__FUNCTION__, L"Message Received");
@@ -276,948 +209,6 @@ namespace winrt::Windows::Devices::Midi2::implementation
         }
     }
 
-    void MidiEndpointConnection::InitializePlugins() noexcept
-    {
-        internal::LogInfo(__FUNCTION__, L"Initializing message processing plugins");
-
-        for (const auto& plugin : m_messageProcessingPlugins)
-        {
-            try
-            {
-                plugin.Initialize(*this);
-            }
-            catch (...)
-            {
-                internal::LogGeneralError(__FUNCTION__, L"Exception initializing plugins.");
-
-            }
-        }
-    }
-
-
-    void MidiEndpointConnection::CallOnConnectionOpenedOnPlugins() noexcept
-    {
-        internal::LogInfo(__FUNCTION__, L"Notifying message processing plugins that the connection is opened");
-
-        for (const auto& plugin : m_messageProcessingPlugins)
-        {
-            try
-            {
-                plugin.OnEndpointConnectionOpened();
-            }
-            catch (...)
-            {
-                internal::LogGeneralError(__FUNCTION__, L"Exception calling Open on plugins.");
-            }
-        }
-    }
-
-    void MidiEndpointConnection::CleanupPlugins() noexcept
-    {
-        internal::LogInfo(__FUNCTION__, L"Cleaning up plugins");
-
-        for (const auto& plugin : m_messageProcessingPlugins)
-        {
-            try
-            {
-                plugin.Cleanup();
-            }
-            catch (...)
-            {
-                internal::LogGeneralError(__FUNCTION__, L"Exception cleaning up plugins.");
-            }
-        }
-    }
-
-
-
-    _Use_decl_annotations_
-    midi2::MidiSendMessageResult MidiEndpointConnection::SendMessageRaw(
-        winrt::com_ptr<IMidiBiDi> endpoint,
-        void* data,
-        uint8_t sizeInBytes,
-        internal::MidiTimestamp timestamp)
-    {
-        internal::LogInfo(__FUNCTION__, L"Sending message raw");
-
-        try
-        {
-            if (!m_isOpen)
-            {
-                internal::LogGeneralError(__FUNCTION__, L"Endpoint is not open. Did you forget to call Open()?");
-
-                // return failure if we're not open
-                return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::EndpointConnectionClosedOrInvalid;
-            }
-
-            if (timestamp != 0 && timestamp > internal::Shared::GetCurrentMidiTimestamp() + m_maxAllowedTimestampOffset)
-            {
-                internal::LogGeneralError(__FUNCTION__, L"Timestamp exceeds maximum future scheduling offset");
-
-                return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::TimestampOutOfRange;
-            }
-
-            if (endpoint != nullptr)
-            {
-                auto hr = endpoint->SendMidiMessage(data, sizeInBytes, timestamp);
-
-                return SendMessageResultFromHRESULT(hr);
-            }
-            else
-            {
-                internal::LogGeneralError(__FUNCTION__, L"endpoint is nullptr");
-
-                return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::EndpointConnectionClosedOrInvalid;
-            }
-        }
-        catch (winrt::hresult_error const& ex)
-        {
-            internal::LogHresultError(__FUNCTION__, L"hresult error sending message. Is the service running?", ex);
-
-            // TOD: Handle other hresults like buffer full
-            return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::Other;
-        }
-    }
-
-
-
-    _Use_decl_annotations_
-    void* MidiEndpointConnection::GetUmpDataPointer(
-        midi2::IMidiUniversalPacket const& ump,
-        uint8_t& dataSizeOut)
-    {
-        void* umpDataPointer{};
-        dataSizeOut = 0;
-
-        switch (ump.PacketType())
-        {
-        case midi2::MidiPacketType::UniversalMidiPacket32:
-            dataSizeOut = (uint8_t)sizeof(internal::PackedUmp32);
-            umpDataPointer = ump.as<implementation::MidiMessage32>()->GetInternalUmpDataPointer();
-            break;
-
-        case midi2::MidiPacketType::UniversalMidiPacket64:
-            dataSizeOut = (uint8_t)sizeof(internal::PackedUmp64);
-            umpDataPointer = ump.as<implementation::MidiMessage64>()->GetInternalUmpDataPointer();
-            break;
-
-        case midi2::MidiPacketType::UniversalMidiPacket96:
-            dataSizeOut = (uint8_t)sizeof(internal::PackedUmp96);
-            umpDataPointer = ump.as<implementation::MidiMessage96>()->GetInternalUmpDataPointer();
-            break;
-
-        case midi2::MidiPacketType::UniversalMidiPacket128:
-            dataSizeOut = (uint8_t)sizeof(internal::PackedUmp128);
-            umpDataPointer = ump.as<implementation::MidiMessage128>()->GetInternalUmpDataPointer();
-            break;
-        }
-
-        return umpDataPointer;
-    }
-
-
-
-    _Use_decl_annotations_
-    midi2::MidiSendMessageResult MidiEndpointConnection::SendMessageStruct(
-        internal::MidiTimestamp timestamp,
-        midi2::MidiMessageStruct const& message,
-        uint8_t wordCount) noexcept
-    {
-        internal::LogInfo(__FUNCTION__, L"Sending message struct");
-
-        if (!ValidateUmp(message.Word0, wordCount))
-        {
-            internal::LogUmpSizeValidationError(__FUNCTION__, L"Word count is incorrect for this UMP", wordCount, timestamp);
-
-            return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::InvalidMessageTypeForWordCount;
-        }
-
-        auto byteLength = (uint8_t)(wordCount * sizeof(uint32_t));
-
-        return SendMessageRaw(m_endpointAbstraction, (void*)(&message), byteLength, timestamp);
-    }
-
-
-
-
-    _Use_decl_annotations_
-    midi2::MidiSendMessageResult MidiEndpointConnection::SendUmpInternal(
-        winrt::com_ptr<IMidiBiDi> endpoint,
-        midi2::IMidiUniversalPacket const& ump)
-    {
-        internal::LogInfo(__FUNCTION__, L"Sending message internal");
-        try
-        {
-            if (endpoint != nullptr)
-            {
-                uint8_t umpDataSize{};
-
-                auto umpDataPointer = GetUmpDataPointer(ump, umpDataSize);
-
-                if (umpDataPointer == nullptr)
-                {
-                    internal::LogGeneralError(__FUNCTION__, L"endpoint data pointer is nullptr");
-
-                    return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::InvalidMessageOther;
-                }
-
-                return SendMessageRaw(endpoint, umpDataPointer, umpDataSize, ump.Timestamp());
-            }
-            else
-            {
-                internal::LogGeneralError(__FUNCTION__, L"endpoint is nullptr");
-
-                return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::EndpointConnectionClosedOrInvalid;
-            }
-        }
-        catch (winrt::hresult_error const& ex)
-        {
-            internal::LogHresultError(__FUNCTION__, L"hresult error sending message. Is the service running?", ex);
-
-            // TODO: handle buffer full and other expected hresults
-            return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::Other;
-        }
-    }
-
-
-    _Use_decl_annotations_
-    midi2::MidiSendMessageResult MidiEndpointConnection::SendMessageBuffer(
-        const internal::MidiTimestamp timestamp,
-        winrt::Windows::Foundation::IMemoryBuffer const& buffer,
-        const uint32_t byteOffset,
-        const uint8_t byteCount) noexcept
-    {
-        internal::LogInfo(__FUNCTION__, L"Sending message buffer");
-
-        try
-        {
-            if (!m_isOpen)
-            {
-                internal::LogGeneralError(__FUNCTION__, L"Endpoint is not open. Did you forget to call Open()?");
-
-                // return failure if we're not open
-                return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::EndpointConnectionClosedOrInvalid;
-            }
-
-            // make sure we're sending only a single UMP
-            uint32_t sizeInWords = byteCount / sizeof(uint32_t);
-
-
-            if (!internal::IsValidSingleUmpWordCount(sizeInWords))
-            {
-                internal::LogUmpSizeValidationError(__FUNCTION__, L"Word count is incorrect for a single UMP", sizeInWords, timestamp);
-
-                //throw hresult_invalid_argument();
-                return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::InvalidMessageTypeForWordCount;
-            }
-
-            winrt::Windows::Foundation::IMemoryBufferReference bufferReference = buffer.CreateReference();
-
-            auto interop = bufferReference.as<IMemoryBufferByteAccess>();
-
-            uint8_t* dataPointer{};
-            uint32_t dataSize{};
-            
-            if (SUCCEEDED(interop->GetBuffer(&dataPointer, &dataSize)))
-            {
-                // make sure we're not going to spin past the end of the buffer
-                if (byteOffset + byteCount > bufferReference.Capacity())
-                {
-                    internal::LogGeneralError(__FUNCTION__, L"Buffer smaller than provided offset + byteLength");
-
-                    return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::DataIndexOutOfRange;
-                }
-
-                // send the ump
-                return SendMessageRaw(m_endpointAbstraction, (void*)(dataPointer + byteOffset), byteCount, timestamp);
-            }
-            else
-            {
-                internal::LogGeneralError(__FUNCTION__, L"Unable to get buffer");
-
-                return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::Other;
-            }
-        }
-        catch (winrt::hresult_error const& ex)
-        {
-            internal::LogHresultError(__FUNCTION__, L"hresult exception sending message", ex);
-
-            // TODO: handle buffer full and other expected hresults
-            return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::Other;
-        }
-    }
-
-    _Use_decl_annotations_
-    midi2::MidiSendMessageResult MidiEndpointConnection::SendMessageWordArray(
-        internal::MidiTimestamp const timestamp,
-        winrt::array_view<uint32_t const> words,
-        uint32_t const startIndex,
-        uint8_t const wordCount
-        ) noexcept
-    {
-        internal::LogInfo(__FUNCTION__, L"Sending message word array");
-
-        try
-        {
-            if (!m_isOpen)
-            {
-                internal::LogGeneralError(__FUNCTION__, L"Endpoint is not open. Did you forget to call Open()?");
-
-                // return failure if we're not open
-                return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::EndpointConnectionClosedOrInvalid;
-            }
-
-            // check for out-of-bounds first
-            if (startIndex + wordCount > words.size())
-            {
-                internal::LogGeneralError(__FUNCTION__, L"Array start index + word count is >= array size");
-
-                return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::DataIndexOutOfRange;
-            }
-
-            // check if the message type is correct for the word count
-            if (wordCount < 1 || !ValidateUmp(words[startIndex], wordCount))
-            {
-                internal::LogUmpSizeValidationError(__FUNCTION__, L"Word count is incorrect for this UMP", wordCount, timestamp);
-
-                return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::InvalidMessageTypeForWordCount;
-            }
-
-            if (m_endpointAbstraction)
-            {
-                auto umpDataSize = (uint8_t)(sizeof(uint32_t) * wordCount);
-
-                // if the service goes down, this will fail
-
-                return SendMessageRaw(m_endpointAbstraction, (void*)(words.data() + startIndex), umpDataSize, timestamp);
-            }
-            else
-            {
-                internal::LogGeneralError(__FUNCTION__, L"Endpoint is nullptr");
-
-                return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::EndpointConnectionClosedOrInvalid;
-            }
-        }
-        catch (winrt::hresult_error const& ex)
-        {
-            internal::LogHresultError(__FUNCTION__, L" hresult exception sending message", ex);
-
-            // TODO: handle buffer full and other expected hresults
-            return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::Other;
-        }
-    }
-
-    _Use_decl_annotations_
-    midi2::MidiSendMessageResult MidiEndpointConnection::SendMessageWords(
-        internal::MidiTimestamp const timestamp,
-        uint32_t const word0) noexcept
-    {
-        internal::LogInfo(__FUNCTION__, L"Sending message words (1)");
-
-        try
-        {
-            if (!m_isOpen)
-            {
-                internal::LogGeneralError(__FUNCTION__, L"Endpoint is not open. Did you forget to call Open()?");
-
-                // return failure if we're not open
-                return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::EndpointConnectionClosedOrInvalid;
-            }
-
-
-            if (internal::GetUmpLengthInMidiWordsFromFirstWord(word0) != UMP32_WORD_COUNT)
-            {
-                // mismatch between the message type and the number of words
-                internal::LogUmpSizeValidationError(__FUNCTION__, L"Word count is incorrect for messageType", UMP32_WORD_COUNT, timestamp);
-                return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::InvalidMessageTypeForWordCount;
-            }
-
-
-            if (m_endpointAbstraction)
-            {
-                auto umpByteCount = (uint8_t)(sizeof(internal::PackedUmp32));
-
-                // if the service goes down, this will fail
-
-                return SendMessageRaw(m_endpointAbstraction, (void*)&word0, umpByteCount, timestamp);
-            }
-            else
-            {
-                internal::LogGeneralError(__FUNCTION__, L"Endpoint is nullptr");
-
-                return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::EndpointConnectionClosedOrInvalid;
-            }
-        }
-        catch (winrt::hresult_error const& ex)
-        {
-            internal::LogHresultError(__FUNCTION__, L" hresult exception sending message", ex);
-
-            // TODO: handle buffer full and other expected hresults
-            return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::Other;
-        }
-    }
-
-
-    _Use_decl_annotations_
-    midi2::MidiSendMessageResult MidiEndpointConnection::SendMessageWords(
-        internal::MidiTimestamp const timestamp,
-        uint32_t const word0,
-        uint32_t const word1) noexcept
-    {
-        internal::LogInfo(__FUNCTION__, L"Sending message words (2)");
-
-        try
-        {
-            if (!m_isOpen)
-            {
-                internal::LogGeneralError(__FUNCTION__, L"Endpoint is not open. Did you forget to call Open()?");
-
-                // return failure if we're not open
-                return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::EndpointConnectionClosedOrInvalid;
-            }
-
-
-            if (internal::GetUmpLengthInMidiWordsFromFirstWord(word0) != UMP64_WORD_COUNT)
-            {
-                // mismatch between the message type and the number of words
-                internal::LogUmpSizeValidationError(__FUNCTION__, L"Word count is incorrect for messageType", UMP64_WORD_COUNT, timestamp);
-                return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::InvalidMessageTypeForWordCount;
-            }
-
-
-            if (m_endpointAbstraction)
-            {
-                auto umpByteCount = (uint8_t)(sizeof(internal::PackedUmp64));
-                internal::PackedUmp64 ump{};
-
-                ump.word0 = word0;
-                ump.word1 = word1;
-
-                // if the service goes down, this will fail
-
-                return SendMessageRaw(m_endpointAbstraction, (void*)&ump, umpByteCount, timestamp);
-            }
-            else
-            {
-                internal::LogGeneralError(__FUNCTION__, L"Endpoint is nullptr");
-
-                return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::EndpointConnectionClosedOrInvalid;
-            }
-        }
-        catch (winrt::hresult_error const& ex)
-        {
-            internal::LogHresultError(__FUNCTION__, L" hresult exception sending message", ex);
-
-            // TODO: handle buffer full and other expected hresults
-            return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::Other;
-        }
-    }
-
-    _Use_decl_annotations_
-    midi2::MidiSendMessageResult MidiEndpointConnection::SendMessageWords(
-        internal::MidiTimestamp const timestamp,
-        uint32_t const word0,
-        uint32_t const word1,
-        uint32_t const word2) noexcept
-    {
-        internal::LogInfo(__FUNCTION__, L"Sending message words (3)");
-
-        try
-        {
-            if (!m_isOpen)
-            {
-                internal::LogGeneralError(__FUNCTION__, L"Endpoint is not open. Did you forget to call Open()?");
-
-                // return failure if we're not open
-                return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::EndpointConnectionClosedOrInvalid;
-            }
-
-
-            if (internal::GetUmpLengthInMidiWordsFromFirstWord(word0) != UMP96_WORD_COUNT)
-            {
-                // mismatch between the message type and the number of words
-                internal::LogUmpSizeValidationError(__FUNCTION__, L"Word count is incorrect for messageType", UMP96_WORD_COUNT, timestamp);
-                return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::InvalidMessageTypeForWordCount;
-            }
-
-
-            if (m_endpointAbstraction)
-            {
-                auto umpByteCount = (uint8_t)(sizeof(internal::PackedUmp96));
-                internal::PackedUmp96 ump{};
-
-                ump.word0 = word0;
-                ump.word1 = word1;
-                ump.word2 = word2;
-
-                // if the service goes down, this will fail
-
-                return SendMessageRaw(m_endpointAbstraction, (void*)&ump, umpByteCount, timestamp);
-            }
-            else
-            {
-                internal::LogGeneralError(__FUNCTION__, L"Endpoint is nullptr");
-
-                return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::EndpointConnectionClosedOrInvalid;
-            }
-        }
-        catch (winrt::hresult_error const& ex)
-        {
-            internal::LogHresultError(__FUNCTION__, L" hresult exception sending message", ex);
-
-            // TODO: handle buffer full and other expected hresults
-            return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::Other;
-        }
-    }
-
-
-    _Use_decl_annotations_
-    midi2::MidiSendMessageResult MidiEndpointConnection::SendMessageWords(
-        internal::MidiTimestamp const timestamp,
-        uint32_t const word0,
-        uint32_t const word1,
-        uint32_t const word2,
-        uint32_t const word3) noexcept
-    {
-        internal::LogInfo(__FUNCTION__, L"Sending message words (4)");
-
-        try
-        {
-            if (!m_isOpen)
-            {
-                internal::LogGeneralError(__FUNCTION__, L"Endpoint is not open. Did you forget to call Open()?");
-
-                // return failure if we're not open
-                return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::EndpointConnectionClosedOrInvalid;
-            }
-
-
-            if (internal::GetUmpLengthInMidiWordsFromFirstWord(word0) != UMP128_WORD_COUNT)
-            {
-                internal::LogUmpSizeValidationError(__FUNCTION__, L"Word count is incorrect for messageType", UMP128_WORD_COUNT, timestamp);
-                return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::InvalidMessageTypeForWordCount;
-            }
-
-
-            if (m_endpointAbstraction)
-            {
-                auto umpByteCount = (uint8_t)(sizeof(internal::PackedUmp128));
-                internal::PackedUmp128 ump{};
-
-                ump.word0 = word0;
-                ump.word1 = word1;
-                ump.word2 = word2;
-                ump.word3 = word3;
-
-                // if the service goes down, this will fail
-
-                return SendMessageRaw(m_endpointAbstraction, (void*)&ump, umpByteCount, timestamp);
-            }
-            else
-            {
-                internal::LogGeneralError(__FUNCTION__, L"Endpoint is nullptr");
-
-                return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::EndpointConnectionClosedOrInvalid;
-            }
-        }
-        catch (winrt::hresult_error const& ex)
-        {
-            internal::LogHresultError(__FUNCTION__, L" hresult exception sending message", ex);
-
-            // TODO: handle buffer full and other expected hresults
-            return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::Other;
-        }
-    }
-
-
-    _Use_decl_annotations_
-    midi2::MidiSendMessageResult MidiEndpointConnection::SendMessagePacket(
-        midi2::IMidiUniversalPacket const& message) noexcept
-    {
-        internal::LogInfo(__FUNCTION__, L"Sending message packet");
-
-        try
-        {
-            if (!m_isOpen)
-            {
-                internal::LogGeneralError(__FUNCTION__, L"Endpoint is not open. Did you forget to call Open()?");
-
-                // return failure if we're not open
-                return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::EndpointConnectionClosedOrInvalid;
-            }
-
-
-            if (m_endpointAbstraction)
-            {
-                return SendUmpInternal(m_endpointAbstraction, message);
-            }
-            else
-            {
-                internal::LogGeneralError(__FUNCTION__, L"Endpoint is nullptr");
-
-                return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::EndpointConnectionClosedOrInvalid;
-            }
-        }
-        catch (winrt::hresult_error const& ex)
-        {
-            internal::LogHresultError(__FUNCTION__, L"hresult exception sending message. Service may be unavailable", ex);
-
-
-            // todo: handle buffer full and similar messages
-            return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::Other;
-        }
-    }
-
-
-
-    _Use_decl_annotations_
-    midi2::MidiSendMessageResult MidiEndpointConnection::SendMultipleMessagesWordList(
-        internal::MidiTimestamp timestamp,
-        collections::IIterable<uint32_t> const& words) noexcept
-    {
-        internal::LogInfo(__FUNCTION__, L"Enter");
-
-        try
-        {
-            if (!m_isOpen)
-            {
-                internal::LogGeneralError(__FUNCTION__, L"Endpoint is not open. Did you forget to call Open()?");
-
-                // return failure if we're not open
-                return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::EndpointConnectionClosedOrInvalid;
-            }
-
-
-            if (m_endpointAbstraction)
-            {
-                // iterate through words based on the message type in the next word
-
-                // we keep reusing this struct
-                internal::PackedUmp128 messageData{};
-
-                //uint32_t index = 0;
-
-                //// TODO: Iteration via index would likely be faster if we just use a array_view, IVector, or IVectorView instead of IIterable.
-
-                //// or maybe we can use a winrt::array_view somehow and get conversion from std?
-                //// https://learn.microsoft.com/en-us/windows/uwp/cpp-and-winrt-apis/std-cpp-data-types
-                //// https://devblogs.microsoft.com/oldnewthing/20200205-00/?p=103398/
-
-                //while (index < totalWords)
-                //{
-
-                //}
-
-
-                auto iter = words.First();
-
-
-                while (iter.HasCurrent())
-                {
-                    auto messageWordCount = internal::GetUmpLengthInMidiWordsFromFirstWord(iter.Current());
-
-                    if (messageWordCount >= 1)
-                    {
-                        messageData.word0 = iter.Current();
-                    }
-                    
-                    if (messageWordCount >= 2)
-                    {
-                        if (iter.MoveNext())
-                        {
-                            messageData.word1 = iter.Current();
-                        }
-                        else
-                        {
-                            internal::LogGeneralError(__FUNCTION__, L"Failed to send message. Message type indicates 2+ words, but we ran out of data.");
-                            return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::DataIndexOutOfRange;
-                        }
-                    }
-                    
-                    if (messageWordCount >= 3)
-                    {
-                        if (iter.MoveNext())
-                        {
-                            messageData.word2 = iter.Current();
-                        }
-                        else
-                        {
-                            internal::LogGeneralError(__FUNCTION__, L"Failed to send message. Message type indicates 3+ words, but we ran out of data.");
-                            return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::DataIndexOutOfRange;
-                        }
-
-                    }
-                    
-                    if (messageWordCount == 4)
-                    {
-                        if (iter.MoveNext())
-                        {
-                            messageData.word3 = iter.Current();
-                        }
-                        else
-                        {
-                            internal::LogGeneralError(__FUNCTION__, L"Failed to send message. Message type indicates 4 words, but we ran out of data.");
-                            return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::DataIndexOutOfRange;
-                        }
-                    }
-
-                    auto sendMessageResult = SendMessageRaw(m_endpointAbstraction, &messageData, messageWordCount * sizeof(uint32_t), timestamp);
-
-                    if (SendMessageFailed(sendMessageResult))
-                    {
-                        // failed. Log and return. We fail on first problem.
-                        internal::LogGeneralError(__FUNCTION__, L"Failed to send message");
-
-                        return sendMessageResult;
-
-                    }
-
-                    iter.MoveNext();
-                }
-
-                return midi2::MidiSendMessageResult::Succeeded;
-
-            }
-            else
-            {
-                internal::LogGeneralError(__FUNCTION__, L"Endpoint is nullptr");
-
-                return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::EndpointConnectionClosedOrInvalid;
-            }
-        }
-        catch (winrt::hresult_error const& ex)
-        {
-            internal::LogHresultError(__FUNCTION__, L"hresult exception sending message. Service may be unavailable", ex);
-
-
-            // todo: handle buffer full and similar messages
-            return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::Other;
-        }
-
-        return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::Other;
-    }
-
-
-
-
-    _Use_decl_annotations_
-    midi2::MidiSendMessageResult MidiEndpointConnection::SendMultipleMessagesBuffer(
-        internal::MidiTimestamp timestamp,
-        foundation::IMemoryBuffer const& buffer,
-        uint32_t byteOffset,
-        uint32_t byteCount) noexcept
-    {
-        internal::LogInfo(__FUNCTION__, L"Enter");
-
-        try
-        {
-            if (!m_isOpen)
-            {
-                internal::LogGeneralError(__FUNCTION__, L"Endpoint is not open. Did you forget to call Open()?");
-
-                // return failure if we're not open
-                return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::EndpointConnectionClosedOrInvalid;
-            }
-
-            winrt::Windows::Foundation::IMemoryBufferReference bufferReference = buffer.CreateReference();
-
-            auto interop = bufferReference.as<IMemoryBufferByteAccess>();
-
-            bool continueProcessing = true;
-            uint8_t* dataPointer{};
-            uint32_t dataSize{ };
-
-            if (SUCCEEDED(interop->GetBuffer(&dataPointer, &dataSize)))
-            {
-                // make sure we're not going to spin past the end of the buffer
-                if (byteOffset + byteCount > bufferReference.Capacity())
-                {
-                    internal::LogGeneralError(__FUNCTION__, L"Buffer smaller than provided offset + byteLength");
-
-                    return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::DataIndexOutOfRange;
-                }
-
-
-                // endianness becomes a concern so we need to make sure we treat this as words
-                //uint32_t* wordDataPointer = (uint32_t*)(dataPointer + byteOffset);
-                uint8_t* byteDataPointer = dataPointer + byteOffset;
-                uint32_t bytesProcessed{ 0 };
-
-                while (continueProcessing)
-                {
-                    if ((byteOffset + bytesProcessed) < dataSize &&
-                        (bytesProcessed + sizeof(uint32_t) <= byteCount))
-                    {
-                        uint8_t bytesInThisMessage = internal::GetUmpLengthInBytesFromFirstWord(*(uint32_t*)byteDataPointer);
-
-                        if (bytesProcessed + bytesInThisMessage <= byteCount)
-                        {
-                            auto result = SendMessageRaw(m_endpointAbstraction, (void*)(byteDataPointer), bytesInThisMessage, timestamp);
-
-                            if (SendMessageFailed(result))
-                            {
-                                internal::LogGeneralError(__FUNCTION__, L"SendMessageRaw failed");
-                                return result;
-                            }
-                        }
-                        else
-                        {
-                            internal::LogGeneralError(__FUNCTION__, L"Ran out of data");
-                            return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::MessageListPartiallyProcessed;
-                        }
-
-                        bytesProcessed += bytesInThisMessage;
-
-                        byteDataPointer += bytesInThisMessage;
-                    }
-                    else
-                    {
-                        continueProcessing = false;
-                    }
-                }
-
-                if (bytesProcessed != byteCount)
-                {
-                    internal::LogGeneralError(__FUNCTION__, L"Ran out of data");
-                    return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::MessageListPartiallyProcessed;
-
-                }
-            }
-            else
-            {
-                internal::LogGeneralError(__FUNCTION__, L"Unable to get buffer");
-
-                return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::Other;
-            }
-
-            return midi2::MidiSendMessageResult::Succeeded;
-        }
-        catch (winrt::hresult_error const& ex)
-        {
-            internal::LogHresultError(__FUNCTION__, L"hresult exception sending message", ex);
-
-            // TODO: handle buffer full and other expected hresults
-            return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::Other;
-        }
-    }
-
-    _Use_decl_annotations_
-    midi2::MidiSendMessageResult MidiEndpointConnection::SendMultipleMessagesStructList(
-        internal::MidiTimestamp timestamp,
-        collections::IIterable<MidiMessageStruct> const& messages) noexcept
-    {
-        internal::LogInfo(__FUNCTION__, L"Enter");
-
-        try
-        {
-            if (!m_isOpen)
-            {
-                internal::LogGeneralError(__FUNCTION__, L"Endpoint is not open. Did you forget to call Open()?");
-
-                // return failure if we're not open
-                return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::EndpointConnectionClosedOrInvalid;
-            }
-
-
-            if (m_endpointAbstraction)
-            {
-                // iterate through words based on the message type in the next word
-
-                auto iter = messages.First();
-
-                for (auto const& message : messages)
-                {
-                    auto messageWordCount = internal::GetUmpLengthInMidiWordsFromFirstWord(message.Word0);
-                    auto sendMessageResult = SendMessageRaw(m_endpointAbstraction, (VOID*)&message, messageWordCount * sizeof(uint32_t), timestamp);
-
-                    if (SendMessageFailed(sendMessageResult))
-                    {
-                        // failed. Log and return. We fail on first problem.
-                        internal::LogGeneralError(__FUNCTION__, L"Failed to send message");
-
-                        return sendMessageResult;
-                    }
-
-
-                }
-
-                return midi2::MidiSendMessageResult::Succeeded;
-
-            }
-            else
-            {
-                internal::LogGeneralError(__FUNCTION__, L"Endpoint is nullptr");
-
-                return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::EndpointConnectionClosedOrInvalid;
-            }
-        }
-        catch (winrt::hresult_error const& ex)
-        {
-            internal::LogHresultError(__FUNCTION__, L"hresult exception sending message. Service may be unavailable", ex);
-
-
-            // todo: handle buffer full and similar messages
-            return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::Other;
-        }
-
-        return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::Other;
-
-    }
-
-
-
-    _Use_decl_annotations_
-    midi2::MidiSendMessageResult MidiEndpointConnection::SendMultipleMessagesPacketList(
-        collections::IIterable<IMidiUniversalPacket> const& messages) noexcept
-    {
-        internal::LogInfo(__FUNCTION__, L"Sending multiple message packet list");
-
-        try
-        {
-            if (!m_isOpen)
-            {
-                internal::LogGeneralError(__FUNCTION__, L"Endpoint is not open. Did you forget to call Open()?");
-
-                // return failure if we're not open
-                return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::EndpointConnectionClosedOrInvalid;
-            }
-
-            if (m_endpointAbstraction)
-            {
-                // right now, we just loop through and send messages. In the future, 
-                // we may optimize this further without changing the API signature
-
-                for (auto const& ump : messages)
-                {
-                    auto result = SendUmpInternal(m_endpointAbstraction, ump);
-
-                    if (!MidiEndpointConnection::SendMessageSucceeded(result))
-                    {
-                        // if any fail, we return immediately.
-
-                        return result;
-                    }
-                }
-
-                return midi2::MidiSendMessageResult::Succeeded;
-            }
-            else
-            {
-                internal::LogGeneralError(__FUNCTION__, L"Endpoint is nullptr");
-
-                return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::EndpointConnectionClosedOrInvalid;
-            }
-        }
-        catch (winrt::hresult_error const& ex)
-        {
-            internal::LogHresultError(__FUNCTION__, L"hresult exception sending messages. Service may be unavailable", ex);
-
-
-            // todo: handle buffer full and similar messages
-            return midi2::MidiSendMessageResult::Failed | midi2::MidiSendMessageResult::Other;
-        }
-    }
-
 
 
 
@@ -1241,44 +232,5 @@ namespace winrt::Windows::Devices::Midi2::implementation
 
             return false;
         }
-    }
-
-
-
-    _Use_decl_annotations_
-    void MidiEndpointConnection::AddMessageProcessingPlugin(midi2::IMidiEndpointMessageProcessingPlugin const& plugin)
-    {
-        m_messageProcessingPlugins.Append(plugin);
-
-        try
-        {
-            plugin.Initialize(*this);
-
-            // if this is added after we've already been opened, call the
-            // handler anyway to get it ready.
-
-            if (m_isOpen)
-            {
-                plugin.OnEndpointConnectionOpened();
-            }
-        }
-        catch (...)
-        {
-            internal::LogGeneralError(__FUNCTION__, L"Exception initializing or calling OnEndpointConnectionOpened on newly-added plugin");
-        }
-    }
-
-    _Use_decl_annotations_
-    void MidiEndpointConnection::RemoveMessageProcessingPlugin(winrt::guid id)
-    {
-        for (uint32_t i = 0; i < m_messageProcessingPlugins.Size(); i++)
-        {
-            if (m_messageProcessingPlugins.GetAt(i).Id() == id)
-            {
-                m_messageProcessingPlugins.RemoveAt(i);
-                break;
-            }
-        }
-
     }
 }
