+++ src/api/Client/Midi2Client/MidiEndpointConnection.idl
@@ -67,7 +67,7 @@ namespace Windows.Devices.Midi2
         // Uses projection-specific array conversion operators for sending data. Not all languages support this
         // Also, arrays are often passed by-val in WinRT, so this may involve a lot of data copying, depending
         // upon implementation. For large buffers without copying, use the IMemoryBuffer
-        MidiSendMessageResult SendMessageWordArray(MIDI_TIMESTAMP timestamp, UInt32[] words, UInt32 startIndex, UInt8 wordCount);
+        MidiSendMessageResult SendMessageWordArray(MIDI_TIMESTAMP timestamp, ref UInt32[] words, UInt32 startIndex, UInt8 wordCount);
 
         // sending words is a very fast approach, but does work with copies of the data. For the most common cases,
         // the amount of data being copied is quite small (32 and 64 bit messages being the most common), and smaller
@@ -82,13 +82,15 @@ namespace Windows.Devices.Midi2
 
         // These methods will send multiple messages. The words must be ordered correctly
         // so they are in order from word-0 to word-n for each message.
-        MidiSendMessageResult SendMultipleMessagesWordList(MIDI_TIMESTAMP timestamp, IIterable<UInt32> words);
-        //MidiSendMessageResult SendMultipleMessagesWordArray(MIDI_TIMESTAMP timestamp, UInt32[] words);
+        MidiSendMessageResult SendMultipleMessagesWordList(MIDI_TIMESTAMP timestamp, IVectorView<UInt32> words);
+        MidiSendMessageResult SendMultipleMessagesWordArray(MIDI_TIMESTAMP timestamp, ref UInt32[] words);
 
         // a convenience for sending multiple packets. We can do this because these are self-describing
-        MidiSendMessageResult SendMultipleMessagesPacketList(IIterable<IMidiUniversalPacket> messages);
-        
-        MidiSendMessageResult SendMultipleMessagesStructList(MIDI_TIMESTAMP timestamp, IIterable<MidiMessageStruct> messages);
+        MidiSendMessageResult SendMultipleMessagesPacketList(IVectorView<IMidiUniversalPacket> messages);
+        MidiSendMessageResult SendMultipleMessagesPacketArray(ref IMidiUniversalPacket[] messages);
+
+        MidiSendMessageResult SendMultipleMessagesStructList(MIDI_TIMESTAMP timestamp, IVectorView<MidiMessageStruct> messages);
+        MidiSendMessageResult SendMultipleMessagesStructArray(MIDI_TIMESTAMP timestamp, ref MidiMessageStruct[] messages);
 
         MidiSendMessageResult SendMultipleMessagesBuffer(MIDI_TIMESTAMP timestamp, Windows.Foundation.IMemoryBuffer buffer, UInt32 byteOffset, UInt32 byteCount);
