+++ src/api/Client/Midi2Client/Windows.Devices.Midi2.vcxproj
@@ -385,6 +385,18 @@
     <ClCompile Include="MidiEndpointConnection.cpp">
       <DependentUpon>MidiEndpointConnection.idl</DependentUpon>
     </ClCompile>
+      <ClCompile Include="MidiEndpointConnection_Plugins.cpp">
+          <DependentUpon>MidiEndpointConnection.idl</DependentUpon>
+      </ClCompile>
+      <ClCompile Include="MidiEndpointConnection_SendInfrastructure.cpp">
+          <DependentUpon>MidiEndpointConnection.idl</DependentUpon>
+      </ClCompile>
+      <ClCompile Include="MidiEndpointConnection_SendMultipleMessages.cpp">
+          <DependentUpon>MidiEndpointConnection.idl</DependentUpon>
+      </ClCompile>
+      <ClCompile Include="MidiEndpointConnection_SendSingleMessage.cpp">
+          <DependentUpon>MidiEndpointConnection.idl</DependentUpon>
+      </ClCompile>
     <ClCompile Include="MidiEndpointDeviceInformation.cpp">
       <DependentUpon>MidiEndpointDeviceInformation.idl</DependentUpon>
     </ClCompile>
